Refactor python tests.
