--- .github/workflows/98-junit-functional-python.yaml
@@ -1,31 +1,0 @@
-# This workflow uses actions that are not certified by GitHub.
-# They are provided by a third-party and are governed by
-# separate terms of service, privacy policy, and support
-# documentation.
-
-name: Java (Functional Tests, Python Integration)
-
-on: [push]
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    timeout-minutes: 25
-
-    steps:
-      - uses: actions/checkout@v2
-      - name: Set up JDK 11
-        uses: actions/setup-java@v2
-        with:
-          java-version: '11'
-          distribution: 'adopt'
-      - name: Validate Gradle wrapper
-        uses: gradle/wrapper-validation-action@e6e38bacfdf1a337459f332974bb2327a31aaf4b
-      - name: Clone Filibuster (Python)
-        run: cd /tmp && git clone http://github.com/filibuster-testing/filibuster -b improvements-2
-      - name: Build Filibuster (Python)
-        run: cd /tmp/filibuster && sudo make
-      - name: Install Filibuster (Python)
-        run: cd /tmp/filibuster && sudo make install
-      - name: Test with Gradle
-        run: ./gradlew testWithFaultInjectionPython
