+++ src/test/java/cloud/filibuster/functional/docker/basic/JUnitFilibusterFailingTest.java
@@ -1,15 +1,17 @@
-package cloud.filibuster.functional.python.exhaustive;
+package cloud.filibuster.functional.docker.basic;
 
 import cloud.filibuster.examples.Hello;
 import cloud.filibuster.examples.HelloServiceGrpc;
+import cloud.filibuster.examples.WorldServiceGrpc;
 import cloud.filibuster.instrumentation.helpers.Networking;
 import cloud.filibuster.junit.TestWithFilibuster;
-import cloud.filibuster.junit.configuration.examples.FilibusterGrpcExhaustiveAnalysisConfigurationFile;
 import cloud.filibuster.junit.interceptors.GitHubActionsSkipInvocationInterceptor;
+import cloud.filibuster.junit.server.backends.FilibusterDockerServerBackend;
 import cloud.filibuster.junit.server.backends.FilibusterLocalProcessServerBackend;
 import cloud.filibuster.functional.JUnitBaseTest;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
+import io.grpc.StatusRuntimeException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
@@ -17,71 +19,77 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
-
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
 import static cloud.filibuster.junit.assertions.protocols.GenericAssertions.wasFaultInjected;
+import static cloud.filibuster.junit.assertions.protocols.GrpcAssertions.wasFaultInjectedOnMethod;
+import static cloud.filibuster.junit.assertions.protocols.GrpcAssertions.wasFaultInjectedOnService;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
+/**
+ * Test simple annotation usage.
+ */
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-@SuppressWarnings("Java8ApiChecker")
-public class JUnitFilibusterTestWithExhaustiveAnalysisFileByAnnotation extends JUnitBaseTest {
-    private static final List<String> exhaustiveGrpcErrorCodeList = new ArrayList<>();
-
-    static {
-        exhaustiveGrpcErrorCodeList.add("CANCELLED");
-        exhaustiveGrpcErrorCodeList.add("UNKNOWN");
-        exhaustiveGrpcErrorCodeList.add("INVALID_ARGUMENT");
-        exhaustiveGrpcErrorCodeList.add("DEADLINE_EXCEEDED");
-        exhaustiveGrpcErrorCodeList.add("NOT_FOUND");
-        exhaustiveGrpcErrorCodeList.add("ALREADY_EXISTS");
-        exhaustiveGrpcErrorCodeList.add("PERMISSION_DENIED");
-        exhaustiveGrpcErrorCodeList.add("RESOURCE_EXHAUSTED");
-        exhaustiveGrpcErrorCodeList.add("FAILED_PRECONDITION");
-        exhaustiveGrpcErrorCodeList.add("ABORTED");
-        exhaustiveGrpcErrorCodeList.add("OUT_OF_RANGE");
-        exhaustiveGrpcErrorCodeList.add("UNIMPLEMENTED");
-        exhaustiveGrpcErrorCodeList.add("INTERNAL");
-        exhaustiveGrpcErrorCodeList.add("UNAVAILABLE");
-        exhaustiveGrpcErrorCodeList.add("DATA_LOSS");
-        exhaustiveGrpcErrorCodeList.add("UNAUTHENTICATED");
-    }
-
+public class JUnitFilibusterFailingTest extends JUnitBaseTest {
     private final static Set<String> testExceptionsThrown = new HashSet<>();
 
+    /**
+     * Inject faults between Hello and World using Filibuster and assert proper faults are injected.
+     *
+     * @throws InterruptedException if teardown of gRPC channel fails.
+     */
     @DisplayName("Test partial hello server grpc route with Filibuster. (MyHelloService, MyWorldService)")
     @ExtendWith(GitHubActionsSkipInvocationInterceptor.class)
-    @TestWithFilibuster(analysisConfigurationFile=FilibusterGrpcExhaustiveAnalysisConfigurationFile.class, serverBackend=FilibusterLocalProcessServerBackend.class)
+    @TestWithFilibuster(serverBackend= FilibusterDockerServerBackend.class, expected=StatusRuntimeException.class)
     @Order(1)
     public void testMyHelloAndMyWorldServiceWithFilibuster() throws InterruptedException {
+        // Intentionally use a bad port.
         ManagedChannel helloChannel = ManagedChannelBuilder
-                .forAddress(Networking.getHost("hello"), Networking.getPort("hello"))
+                .forAddress(Networking.getHost("hello"), 9999)
                 .usePlaintext()
                 .build();
+
+        boolean expected = false;
 
         try {
             HelloServiceGrpc.HelloServiceBlockingStub blockingStub = HelloServiceGrpc.newBlockingStub(helloChannel);
             Hello.HelloRequest request = Hello.HelloRequest.newBuilder().setName("Armerian").build();
             Hello.HelloReply reply = blockingStub.partialHello(request);
             assertEquals("Hello, Armerian World!!", reply.getMessage());
+            assertFalse(wasFaultInjected());
         } catch (Throwable t) {
-            if (wasFaultInjected()) {
+            boolean wasFaultInjected = wasFaultInjected();
+
+            if (wasFaultInjected) {
                 testExceptionsThrown.add(t.getMessage());
 
-                boolean found = false;
-
-                for (String errorCode: exhaustiveGrpcErrorCodeList) {
-                    String expectedString = "DATA_LOSS: io.grpc.StatusRuntimeException: " + errorCode;
-                    if(t.getMessage().equals(expectedString)) {
-                        found = true;
-                    }
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: DEADLINE_EXCEEDED")) {
+                    expected = true;
                 }
 
-                if (! found) {
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNAVAILABLE")) {
+                    expected = true;
+                }
+
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNIMPLEMENTED")) {
+                    expected = true;
+                }
+
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: INTERNAL")) {
+                    expected = true;
+                }
+
+                boolean wasFaultInjectedOnWorldService = wasFaultInjectedOnService("World");
+                assertTrue(wasFaultInjectedOnWorldService);
+
+                boolean wasFaultInjectedOnWorldMethod = wasFaultInjectedOnMethod(WorldServiceGrpc.getWorldMethod());
+                assertTrue(wasFaultInjectedOnWorldMethod);
+
+                if (! expected) {
                     throw t;
                 }
             } else {
@@ -93,11 +101,14 @@ public class JUnitFilibusterTestWithExhaustiveAnalysisFileByAnnotation extends J
         helloChannel.awaitTermination(1000, TimeUnit.SECONDS);
     }
 
+    /**
+     * Verify that Filibuster generated the correct number of fault injections.
+     */
     @DisplayName("Verify correct number of generated Filibuster tests.")
     @ExtendWith(GitHubActionsSkipInvocationInterceptor.class)
     @Test
     @Order(2)
     public void testNumAssertions() {
-        assertEquals(16, testExceptionsThrown.size());
+        assertEquals(0, testExceptionsThrown.size());
     }
 }
