+++ src/test/java/cloud/filibuster/functional/docker/basic/JUnitFilibusterTest.java
@@ -1,10 +1,11 @@
-package cloud.filibuster.functional.python.extended;
+package cloud.filibuster.functional.docker.basic;
 
 import cloud.filibuster.examples.Hello;
 import cloud.filibuster.examples.HelloServiceGrpc;
+import cloud.filibuster.examples.WorldServiceGrpc;
 import cloud.filibuster.instrumentation.helpers.Networking;
 import cloud.filibuster.junit.TestWithFilibuster;
-import cloud.filibuster.junit.interceptors.GitHubActionsSkipInvocationInterceptor;
+import cloud.filibuster.junit.server.backends.FilibusterDockerServerBackend;
 import cloud.filibuster.junit.server.backends.FilibusterLocalProcessServerBackend;
 import cloud.filibuster.functional.JUnitBaseTest;
 import io.grpc.ManagedChannel;
@@ -14,35 +15,32 @@ import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
-import org.junit.jupiter.api.extension.ExtendWith;
 
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
 import static cloud.filibuster.junit.assertions.protocols.GenericAssertions.wasFaultInjected;
+import static cloud.filibuster.junit.assertions.protocols.GrpcAssertions.wasFaultInjectedOnMethod;
+import static cloud.filibuster.junit.assertions.protocols.GrpcAssertions.wasFaultInjectedOnService;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
+/**
+ * Test simple annotation usage.
+ */
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-@SuppressWarnings("Java8ApiChecker")
-public class JUnitFilibusterTestWithBasicAnalysisFileByAnnotation extends JUnitBaseTest {
-    private static final List<String> basicGrpcErrorCodeList = new ArrayList<>();
-
-    static {
-        basicGrpcErrorCodeList.add("DEADLINE_EXCEEDED");
-        basicGrpcErrorCodeList.add("UNAVAILABLE");
-        basicGrpcErrorCodeList.add("INTERNAL");
-        basicGrpcErrorCodeList.add("UNIMPLEMENTED");
-        basicGrpcErrorCodeList.add("UNKNOWN");
-    }
-
+public class JUnitFilibusterTest extends JUnitBaseTest {
     private final static Set<String> testExceptionsThrown = new HashSet<>();
 
+    /**
+     * Inject faults between Hello and World using Filibuster and assert proper faults are injected.
+     *
+     * @throws InterruptedException if teardown of gRPC channel fails.
+     */
     @DisplayName("Test partial hello server grpc route with Filibuster. (MyHelloService, MyWorldService)")
-    @ExtendWith(GitHubActionsSkipInvocationInterceptor.class)
-    @TestWithFilibuster(serverBackend=FilibusterLocalProcessServerBackend.class)
+    @TestWithFilibuster(serverBackend= FilibusterDockerServerBackend.class)
     @Order(1)
     public void testMyHelloAndMyWorldServiceWithFilibuster() throws InterruptedException {
         ManagedChannel helloChannel = ManagedChannelBuilder
@@ -50,25 +48,47 @@ public class JUnitFilibusterTestWithBasicAnalysisFileByAnnotation extends JUnitB
                 .usePlaintext()
                 .build();
 
+        boolean expected = false;
+
         try {
             HelloServiceGrpc.HelloServiceBlockingStub blockingStub = HelloServiceGrpc.newBlockingStub(helloChannel);
             Hello.HelloRequest request = Hello.HelloRequest.newBuilder().setName("Armerian").build();
             Hello.HelloReply reply = blockingStub.partialHello(request);
             assertEquals("Hello, Armerian World!!", reply.getMessage());
+            assertFalse(wasFaultInjected());
         } catch (Throwable t) {
-            if (wasFaultInjected()) {
+            boolean wasFaultInjected = wasFaultInjected();
+
+            if (wasFaultInjected) {
                 testExceptionsThrown.add(t.getMessage());
 
-                boolean found = false;
-
-                for (String errorCode: basicGrpcErrorCodeList) {
-                    String expectedString = "DATA_LOSS: io.grpc.StatusRuntimeException: " + errorCode;
-                    if(t.getMessage().equals(expectedString)) {
-                        found = true;
-                    }
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: DEADLINE_EXCEEDED")) {
+                    expected = true;
                 }
 
-                if (! found) {
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNAVAILABLE")) {
+                    expected = true;
+                }
+
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNIMPLEMENTED")) {
+                    expected = true;
+                }
+
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: INTERNAL")) {
+                    expected = true;
+                }
+
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNKNOWN")) {
+                    expected = true;
+                }
+
+                boolean wasFaultInjectedOnWorldService = wasFaultInjectedOnService("world");
+                assertTrue(wasFaultInjectedOnWorldService);
+
+                boolean wasFaultInjectedOnWorldMethod = wasFaultInjectedOnMethod(WorldServiceGrpc.getWorldMethod());
+                assertTrue(wasFaultInjectedOnWorldMethod);
+
+                if (! expected) {
                     throw t;
                 }
             } else {
@@ -80,8 +100,10 @@ public class JUnitFilibusterTestWithBasicAnalysisFileByAnnotation extends JUnitB
         helloChannel.awaitTermination(1000, TimeUnit.SECONDS);
     }
 
+    /**
+     * Verify that Filibuster generated the correct number of fault injections.
+     */
     @DisplayName("Verify correct number of generated Filibuster tests.")
-    @ExtendWith(GitHubActionsSkipInvocationInterceptor.class)
     @Test
     @Order(2)
     public void testNumAssertions() {
