+++ src/test/java/cloud/filibuster/functional/docker/custom/JUnitFilibusterTestWithCustomAnalysisFile.java
@@ -1,4 +1,4 @@
-package cloud.filibuster.functional.python.armeria;
+package cloud.filibuster.functional.docker.custom;
 
 import cloud.filibuster.examples.Hello;
 import cloud.filibuster.examples.HelloServiceGrpc;
@@ -7,11 +7,11 @@ import cloud.filibuster.junit.TestWithFilibuster;
 import cloud.filibuster.junit.configuration.FilibusterAnalysisConfiguration;
 import cloud.filibuster.junit.configuration.FilibusterCustomAnalysisConfigurationFile;
 import cloud.filibuster.junit.interceptors.GitHubActionsSkipInvocationInterceptor;
+import cloud.filibuster.junit.server.backends.FilibusterDockerServerBackend;
 import cloud.filibuster.junit.server.backends.FilibusterLocalProcessServerBackend;
 import cloud.filibuster.functional.JUnitBaseTest;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
@@ -26,23 +26,20 @@ import java.util.concurrent.TimeUnit;
 
 import static cloud.filibuster.junit.assertions.protocols.GenericAssertions.wasFaultInjected;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * Verify that the Filibuster analysis can be configured to use a custom set of faults.
  */
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 @SuppressWarnings("Java8ApiChecker")
-public class JUnitFilibusterTestWithFailFastException extends JUnitBaseTest {
+public class JUnitFilibusterTestWithCustomAnalysisFile extends JUnitBaseTest {
     private static final String analysisFilePath = "/tmp/filibuster-custom-analysis-file";
 
-    @BeforeAll
-    public static void writeCustomAnalysisFile() {
+    static {
+        @SuppressWarnings("Java8ApiChecker")
         FilibusterAnalysisConfiguration filibusterAnalysisConfiguration = new FilibusterAnalysisConfiguration.Builder()
                 .name("java.grpc")
                 .pattern("(.*/.*)")
-
-                // Base exceptions.
                 .exception("io.grpc.StatusRuntimeException", Map.of(
                         "cause", "",
                         "code", "UNAVAILABLE"
@@ -59,23 +56,11 @@ public class JUnitFilibusterTestWithFailFastException extends JUnitBaseTest {
                         "cause", "",
                         "code", "NOT_FOUND"
                 ))
-                .exception("io.grpc.StatusRuntimeException", Map.of(
-                        "cause", "",
-                        "code", "UNKNOWN"
-                ))
-
-                // Custom exception type, but must support the RuntimeException set of constructors.
-                .exception("io.grpc.StatusRuntimeException", Map.of(
-                        "cause", "cloud.filibuster.exceptions.CircuitBreakerException",
-                        "code", "UNKNOWN"
-                ))
-
                 .build();
         FilibusterCustomAnalysisConfigurationFile filibusterAnalysisConfigurationFile = new FilibusterCustomAnalysisConfigurationFile.Builder()
                 .analysisConfiguration(filibusterAnalysisConfiguration)
                 .build();
-        boolean result = filibusterAnalysisConfigurationFile.writeToDisk(analysisFilePath);
-        assertTrue(result);
+        filibusterAnalysisConfigurationFile.writeToDisk(analysisFilePath);
     }
 
     private final static Set<String> testExceptionsThrown = new HashSet<>();
@@ -86,7 +71,7 @@ public class JUnitFilibusterTestWithFailFastException extends JUnitBaseTest {
      * @throws InterruptedException thrown if the gRPC channel fails to terminate.
      */
     @DisplayName("Test partial hello server grpc route with Filibuster. (MyHelloService, MyWorldService)")
-    @TestWithFilibuster(analysisFile=analysisFilePath, serverBackend=FilibusterLocalProcessServerBackend.class)
+    @TestWithFilibuster(analysisFile=analysisFilePath, serverBackend= FilibusterDockerServerBackend.class)
     @ExtendWith(GitHubActionsSkipInvocationInterceptor.class)
     @Order(1)
     public void testMyHelloAndMyWorldServiceWithFilibuster() throws InterruptedException {
@@ -122,14 +107,6 @@ public class JUnitFilibusterTestWithFailFastException extends JUnitBaseTest {
                     expected = true;
                 }
 
-                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNKNOWN")) {
-                    expected = true;
-                }
-
-                if (t.getMessage().equals("INTERNAL: io.grpc.StatusRuntimeException: UNKNOWN")) {
-                    expected = true;
-                }
-
                 if (! expected) {
                     throw t;
                 }
@@ -150,6 +127,6 @@ public class JUnitFilibusterTestWithFailFastException extends JUnitBaseTest {
     @Test
     @Order(2)
     public void testNumAssertions() {
-        assertEquals(6, testExceptionsThrown.size());
+        assertEquals(4, testExceptionsThrown.size());
     }
 }
