+++ src/test/java/cloud/filibuster/functional/docker/exhaustive/JUnitFilibusterTestWithExhaustiveAnalysisFile.java
@@ -1,4 +1,4 @@
-package cloud.filibuster.functional.python.exhaustive;
+package cloud.filibuster.functional.docker.exhaustive;
 
 import cloud.filibuster.examples.Hello;
 import cloud.filibuster.examples.HelloServiceGrpc;
@@ -7,6 +7,7 @@ import cloud.filibuster.junit.TestWithFilibuster;
 import cloud.filibuster.junit.configuration.FilibusterAnalysisConfiguration;
 import cloud.filibuster.junit.configuration.FilibusterCustomAnalysisConfigurationFile;
 import cloud.filibuster.junit.interceptors.GitHubActionsSkipInvocationInterceptor;
+import cloud.filibuster.junit.server.backends.FilibusterDockerServerBackend;
 import cloud.filibuster.junit.server.backends.FilibusterLocalProcessServerBackend;
 import cloud.filibuster.functional.JUnitBaseTest;
 import io.grpc.ManagedChannel;
@@ -30,7 +31,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 @SuppressWarnings("Java8ApiChecker")
-public class JUnitFilibusterTestWithExhaustiveAnalysisFileMaxIterationsNotEnough extends JUnitBaseTest {
+public class JUnitFilibusterTestWithExhaustiveAnalysisFile extends JUnitBaseTest {
     private static final String analysisFilePath = "/tmp/filibuster-exhaustive-analysis-file";
     private static final List<String> exhaustiveGrpcErrorCodeList = new ArrayList<>();
 
@@ -74,7 +75,7 @@ public class JUnitFilibusterTestWithExhaustiveAnalysisFileMaxIterationsNotEnough
 
     @DisplayName("Test partial hello server grpc route with Filibuster. (MyHelloService, MyWorldService)")
     @ExtendWith(GitHubActionsSkipInvocationInterceptor.class)
-    @TestWithFilibuster(analysisFile=analysisFilePath, maxIterations=3, serverBackend=FilibusterLocalProcessServerBackend.class)
+    @TestWithFilibuster(analysisFile=analysisFilePath, serverBackend= FilibusterDockerServerBackend.class)
     @Order(1)
     public void testMyHelloAndMyWorldServiceWithFilibuster() throws InterruptedException {
         ManagedChannel helloChannel = ManagedChannelBuilder
@@ -117,6 +118,6 @@ public class JUnitFilibusterTestWithExhaustiveAnalysisFileMaxIterationsNotEnough
     @Test
     @Order(2)
     public void testNumAssertions() {
-        assertEquals(2, testExceptionsThrown.size());
+        assertEquals(16, testExceptionsThrown.size());
     }
 }
