+++ src/test/java/cloud/filibuster/functional/docker/macros/JUnitFilibusterTestMacroAssertion.java
@@ -1,11 +1,11 @@
-package cloud.filibuster.functional.python.macros;
+package cloud.filibuster.functional.docker.macros;
 
 import cloud.filibuster.examples.Hello;
 import cloud.filibuster.examples.HelloServiceGrpc;
-import cloud.filibuster.examples.WorldServiceGrpc;
 import cloud.filibuster.instrumentation.helpers.Networking;
 import cloud.filibuster.junit.TestWithFilibuster;
 import cloud.filibuster.junit.interceptors.GitHubActionsSkipInvocationInterceptor;
+import cloud.filibuster.junit.server.backends.FilibusterDockerServerBackend;
 import cloud.filibuster.junit.server.backends.FilibusterLocalProcessServerBackend;
 import cloud.filibuster.functional.JUnitBaseTest;
 import io.grpc.ManagedChannel;
@@ -15,12 +15,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.opentest4j.AssertionFailedError;
 
 import java.util.concurrent.TimeUnit;
 
-import static cloud.filibuster.junit.assertions.protocols.GrpcAssertions.wasFaultInjectedOnService;
-import static cloud.filibuster.junit.assertions.protocols.GrpcAssertions.wasFaultInjectedOnMethod;
 import static cloud.filibuster.junit.assertions.protocols.GrpcAssertions.tryGrpcAndCatchGrpcExceptions;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -29,11 +26,11 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  * Test simple annotation usage.
  */
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitFilibusterTestExtendedMacroAssertion extends JUnitBaseTest {
+public class JUnitFilibusterTestMacroAssertion extends JUnitBaseTest {
 
     @DisplayName("Test partial hello server grpc route with Filibuster. (MyHelloService, MyWorldService)")
     @ExtendWith(GitHubActionsSkipInvocationInterceptor.class)
-    @TestWithFilibuster(serverBackend=FilibusterLocalProcessServerBackend.class)
+    @TestWithFilibuster(serverBackend= FilibusterDockerServerBackend.class)
     public void testMyHelloAndMyWorldServiceWithFilibusterWithMacro() throws Throwable {
         ManagedChannel helloChannel = ManagedChannelBuilder
                 .forAddress(Networking.getHost("hello"), Networking.getPort("hello"))
@@ -47,38 +44,6 @@ public class JUnitFilibusterTestExtendedMacroAssertion extends JUnitBaseTest {
             assertEquals("Hello, Armerian World!!", reply.getMessage());
         }, (t) -> {
             assertTrue(t instanceof StatusRuntimeException);
-
-            boolean expected = false;
-
-            if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: DEADLINE_EXCEEDED")) {
-                expected = true;
-            }
-
-            if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNAVAILABLE")) {
-                expected = true;
-            }
-
-            if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNIMPLEMENTED")) {
-                expected = true;
-            }
-
-            if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: INTERNAL")) {
-                expected = true;
-            }
-
-            if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNKNOWN")) {
-                expected = true;
-            }
-
-            boolean wasFaultInjectedOnWorldService = wasFaultInjectedOnService("world");
-            assertTrue(wasFaultInjectedOnWorldService);
-
-            boolean wasFaultInjectedOnWorldMethod = wasFaultInjectedOnMethod(WorldServiceGrpc.getWorldMethod());
-            assertTrue(wasFaultInjectedOnWorldMethod);
-
-            if (! expected) {
-                throw new AssertionFailedError("Received unexpected exception.", t);
-            }
         });
 
         helloChannel.shutdownNow();
@@ -100,7 +65,7 @@ public class JUnitFilibusterTestExtendedMacroAssertion extends JUnitBaseTest {
             Hello.HelloReply reply = blockingStub.partialHello(request);
             assertEquals("Hello, Armerian World!!", reply.getMessage());
         }, (t) -> {
-            assertEquals(true, false); // Should never get here since we never get past the initial execution.
+            assertTrue(t instanceof StatusRuntimeException);
         });
 
         helloChannel.shutdownNow();
