--- src/test/java/cloud/filibuster/functional/python/basic/JUnitFilibusterFailingTest.java
@@ -1,113 +1,0 @@
-package cloud.filibuster.functional.python.basic;
-
-import cloud.filibuster.examples.Hello;
-import cloud.filibuster.examples.HelloServiceGrpc;
-import cloud.filibuster.examples.WorldServiceGrpc;
-import cloud.filibuster.instrumentation.helpers.Networking;
-import cloud.filibuster.junit.TestWithFilibuster;
-import cloud.filibuster.junit.interceptors.GitHubActionsSkipInvocationInterceptor;
-import cloud.filibuster.junit.server.backends.FilibusterLocalProcessServerBackend;
-import cloud.filibuster.functional.JUnitBaseTest;
-import io.grpc.ManagedChannel;
-import io.grpc.ManagedChannelBuilder;
-import io.grpc.StatusRuntimeException;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.MethodOrderer;
-import org.junit.jupiter.api.Order;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestMethodOrder;
-import org.junit.jupiter.api.extension.ExtendWith;
-
-import java.util.HashSet;
-import java.util.Set;
-import java.util.concurrent.TimeUnit;
-
-import static cloud.filibuster.junit.assertions.protocols.GenericAssertions.wasFaultInjected;
-import static cloud.filibuster.junit.assertions.protocols.GrpcAssertions.wasFaultInjectedOnMethod;
-import static cloud.filibuster.junit.assertions.protocols.GrpcAssertions.wasFaultInjectedOnService;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-/**
- * Test simple annotation usage.
- */
-@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitFilibusterFailingTest extends JUnitBaseTest {
-    private final static Set<String> testExceptionsThrown = new HashSet<>();
-
-    /**
-     * Inject faults between Hello and World using Filibuster and assert proper faults are injected.
-     *
-     * @throws InterruptedException if teardown of gRPC channel fails.
-     */
-    @DisplayName("Test partial hello server grpc route with Filibuster. (MyHelloService, MyWorldService)")
-    @ExtendWith(GitHubActionsSkipInvocationInterceptor.class)
-    @TestWithFilibuster(serverBackend=FilibusterLocalProcessServerBackend.class, expected=StatusRuntimeException.class)
-    @Order(1)
-    public void testMyHelloAndMyWorldServiceWithFilibuster() throws InterruptedException {
-        // Intentionally use a bad port.
-        ManagedChannel helloChannel = ManagedChannelBuilder
-                .forAddress(Networking.getHost("hello"), 9999)
-                .usePlaintext()
-                .build();
-
-        boolean expected = false;
-
-        try {
-            HelloServiceGrpc.HelloServiceBlockingStub blockingStub = HelloServiceGrpc.newBlockingStub(helloChannel);
-            Hello.HelloRequest request = Hello.HelloRequest.newBuilder().setName("Armerian").build();
-            Hello.HelloReply reply = blockingStub.partialHello(request);
-            assertEquals("Hello, Armerian World!!", reply.getMessage());
-            assertFalse(wasFaultInjected());
-        } catch (Throwable t) {
-            boolean wasFaultInjected = wasFaultInjected();
-
-            if (wasFaultInjected) {
-                testExceptionsThrown.add(t.getMessage());
-
-                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: DEADLINE_EXCEEDED")) {
-                    expected = true;
-                }
-
-                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNAVAILABLE")) {
-                    expected = true;
-                }
-
-                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNIMPLEMENTED")) {
-                    expected = true;
-                }
-
-                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: INTERNAL")) {
-                    expected = true;
-                }
-
-                boolean wasFaultInjectedOnWorldService = wasFaultInjectedOnService("World");
-                assertTrue(wasFaultInjectedOnWorldService);
-
-                boolean wasFaultInjectedOnWorldMethod = wasFaultInjectedOnMethod(WorldServiceGrpc.getWorldMethod());
-                assertTrue(wasFaultInjectedOnWorldMethod);
-
-                if (! expected) {
-                    throw t;
-                }
-            } else {
-                throw t;
-            }
-        }
-
-        helloChannel.shutdownNow();
-        helloChannel.awaitTermination(1000, TimeUnit.SECONDS);
-    }
-
-    /**
-     * Verify that Filibuster generated the correct number of fault injections.
-     */
-    @DisplayName("Verify correct number of generated Filibuster tests.")
-    @ExtendWith(GitHubActionsSkipInvocationInterceptor.class)
-    @Test
-    @Order(2)
-    public void testNumAssertions() {
-        assertEquals(0, testExceptionsThrown.size());
-    }
-}
